pyproject.toml
[project]
name = "kr-health-monitor"
version = "0.1.0"
requires-python = ">=3.11"
description = "Public infrastructure for monitoring NHIS coverage adequacy gaps in Korean medical device reimbursement"
authors = [{name = "pon00050"}]
license = {text = "MIT"}
readme = "README.md"
keywords = ["nhis", "hira", "coverage-gap", "cgm", "korea", "open-data", "health-policy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "anthropic>=0.40.0",
    "pandas",
    "pyarrow",
    "duckdb>=1.0.0",
    "requests",
    "python-dotenv",
    "typer>=0.9.0",
    "tqdm",
    "marimo",
    "plotly",
    "lxml",
    "pydantic>=2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "cachetools>=5.3.0",
    "fastmcp>=3.1.0",
]

[project.scripts]
krh = "cli:app"

[project.urls]
Repository = "https://github.com/pon00050/kr-health-monitor"
Issues = "https://github.com/pon00050/kr-health-monitor/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["02_Pipeline", "03_Analysis", "src"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.24.0",
    "pytest-asyncio>=0.23.0",
]

[tool.uv]
compile-bytecode = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
